web_service_monitoring/prediction_service/requirements.txt
Werkzeug~=2.0.1
prometheus_client~=0.11.0
pyyaml~=5.4.1
evidently
pymongo
flask
gunicorn
mlflow
requests
sklearn
xgboost
numpy
python-dotenv
boto3
